# src/lib/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Hindus R Us",
    "description": "An API documentation generated by swagger-autogen."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://h-dummy-app-express-app-182702284982.asia-south1.run.app/"
    }
  ],
  "paths": {
...
    "/v1/users/create": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "An API route to create users on Firebase",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": true,
                  "message": "Ok",
                  "errorCode": "200"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": true,
                  "message": "Bad Request",
                  "errorCode": "400"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBodyRequest"
              }
            }
          }
        }
      }
    },
    "/v1/users/generateId": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Fetch User Details",
        "description": "An API route to fetch user details from Identity platform",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": true,
                  "message": "Ok",
                  "errorCode": "200"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": true,
                  "message": "Bad Request",
                  "errorCode": "400"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdGenSchema"
              }
            }
          }
        }
      }
    },
...
    "/v1/users/delete": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "IdGenSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "returnSecureToken": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "password",
          "returnSecureToken"
        ]
      },
...
      "CreateBodyRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@example.com"
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "phoneNumber": {
            "type": "string",
            "example": "+91 1231231231"
          },
          "displayName": {
            "type": "string",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "email",
          "displayName",
          "password",
          "role"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}